refactor: split main js file into smaller modules
Broke the single JS file into 5 smaller files.
formSubmit is called on the submission of the travelInfo form.
formSubmit then calls formValidation to ensure propery use of form fields.
formSubmit then calls getLocationInfo to fetch GeoNames API on the
destination. The data from the GeoNames API is then sent to postData to send
the information to the server.