feat: Add the handling of weather information
Add logic to to check if date of trip is in 16 day forcast range.
Use the logic to give starting date for weather output.
Use a for loop to go through weather data and generate DOM elements for
each day that weather forcast is available.
Append the elements to DOM.